+++ Source/Injector/Pegasus.cs
@@ -15,7 +15,7 @@ using System.Windows.Forms;
 
 internal static class Program {
     internal const string DllName = "BedrockUtilityFramework.Xray.dll";
-    internal const string ExpectedHash = "E08CA009BCBF5A8C2459CEC39571647D66035EA773DA757FD70F950D444AD5E6";
+    internal const string ExpectedHash = "B2E1BBDF39DCFCFD5BDE3FA185C56C002DB451012C586F366EBB825B7C35E2B1";
     internal static string DllPath { get { return Path.Combine(AppDomain.CurrentDomain.BaseDirectory, DllName); } }
     [STAThread] static int Main(string[] args) {
         if (args.Length == 1 && args[0] == "--test-host") { Thread.Sleep(30000); return 0; }
@@ -235,3 +235,4 @@ internal static class Loader {
         }
     }
 }
+