+++ Source/Mod/src/framework/Framework.cpp
@@ -125,3 +125,4 @@ bool Framework::initialized() const noexcept {
 }
 
 } // namespace utility
+