+++ Source/Mod/src/modules/ReachModule.cpp
@@ -619,3 +619,4 @@ void ReachModule::uninstall_hooks() noexcept {
 }
 
 } // namespace utility::modules
+