+++ Source/Mod/src/modules/ReachModule.hpp
@@ -82,3 +82,4 @@ private:
     ReachModule& entity_;
 };
 } // namespace utility::modules
+